devtools/server/tests/mochitest/test_framerate_01.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1007200 - Create a framerate actor
-->
<head>
  <meta charset="utf-8">
  <title>Framerate actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = function () {
  const Cu = Components.utils;

  const { require } = Cu.import("resource://devtools/shared/Loader.jsm", {});
  const Services = require("Services");
  const { DebuggerClient } = require("devtools/shared/client/main");
  const { DebuggerServer } = require("devtools/server/main");

  // Always log packets when running tests.
  Services.prefs.setBoolPref("devtools.debugger.log", true);
  SimpleTest.registerCleanupFunction(function () {
    Services.prefs.clearUserPref("devtools.debugger.log");
  });

  SimpleTest.waitForExplicitFinish();

  const { FramerateFront } = require("devtools/shared/fronts/framerate");

  function plotFPS(ticks, interval = 100, clamp = 60) {
    let timeline = [];
    let totalTicks = ticks.length;

    // If the refresh driver didn't get a chance to tick before the
    // recording was stopped, assume framerate was 0.
    if (totalTicks == 0) {
      timeline.push({ delta: 0, value: 0 });
      timeline.push({ delta: interval, value: 0 });
      return timeline;
    }

    let frameCount = 0;
    let prevTime = ticks[0];

    for (let i = 1; i < totalTicks; i++) {
      let currTime = ticks[i];
      frameCount++;

      let elapsedTime = currTime - prevTime;
      if (elapsedTime < interval) {
        continue;
      }

      let framerate = Math.min(1000 / (elapsedTime / frameCount), clamp);
      timeline.push({ delta: prevTime, value: framerate });
      timeline.push({ delta: currTime, value: framerate });

      frameCount = 0;
      prevTime = currTime;
    }

    return timeline;
  }

  if (!DebuggerServer.initialized) {
    DebuggerServer.init();
    DebuggerServer.addBrowserActors();
  }

  let client = new DebuggerClient(DebuggerServer.connectPipe());
  client.connect().then(function onConnect() {
    client.listTabs(function onListTabs(response) {
      let form = response.tabs[response.selected];
      let front = FramerateFront(client, form);

      window.setTimeout(() => {
        /* eslint-disable max-nested-callbacks */
        front.startRecording().then(() => {
          window.setTimeout(() => {
            front.stopRecording().then(rawData => {
              onRecordingStopped(front, rawData);
            });
          }, 1000);
        });
        /* eslint-enable max-nested-callbacks */
      }, 1000);
    });
  });

  function onRecordingStopped(front, rawData) {
    ok(rawData, "There should be a recording available.");

    let timeline = plotFPS(rawData);
    ok(timeline.length >= 2,
      "There should be at least one measurement available, with two entries.");

    let prevTimeStart = timeline[0].delta;

    for (let i = 0; i < timeline.length; i += 2) {
      let currTimeStart = timeline[i].delta;
      let currTimeEnd = timeline[i + 1].delta;
      info("Testing delta: " + currTimeStart + " vs. " + currTimeEnd);

      ok(currTimeStart < currTimeEnd,
        "The start and end time deltas should be consecutive.");
      is(currTimeStart, prevTimeStart,
        "There should be two time deltas for each framerate value.");

      prevTimeStart = currTimeEnd;
    }

    for (let i = 0; i < timeline.length; i += 2) {
      let currFramerateStart = timeline[i].value;
      let currFramerateEnd = timeline[i + 1].value;
      info("Testing framerate: " + currFramerateStart);

      is(currFramerateStart, currFramerateEnd,
        "The start and end framerate values should be equal.");

      is(typeof currFramerateStart, "number", "All values should be numbers.");
      ok(currFramerateStart <= 60, "All values were correctly clamped.");
    }

    client.close().then(() => {
      DebuggerServer.destroy();
      SimpleTest.finish();
    });
  }
};
</script>
</pre>
</body>
</html>
